pyproject.toml
[tool.poetry]
name = "data-field-diff"
version = "0.1.0"
description = ""
authors = ["Dan <dntknwhtf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.22.0"
tabulate = "^0.9.0"
pandas = "^2.2.2"
datetime = "^5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
